Refactored UserArea components
